+++ changelog.md
@@ -5,11 +5,11 @@
 * Per-entity profiling (mean, max, ...)
 * Client sync by chat with interaction link
   * Optional on login, otherwise just on changes by command
-* Use distant culling with shouldRenderAtSqrDistance
 
 # 0.0.1
 
 * Added basic tick scheduler and config
   * Viewport culling
+  * Distance culling
   * Client linking on integrated servers
 * Server profile report and configuration commands